feat(NotificationSettingChangeCreate): add validate target list is empty

+++ src/main/java/com/dongsoop/dongsoop/notification/setting/service/handler/NotificationSettingChangeDelete.java
@@ -27,6 +27,10 @@ public class NotificationSettingChangeDelete implements NotificationSettingChang
                 .filter(Objects::nonNull)
                 .toList();
 
+        if (notificationSettingList.isEmpty()) {
+            return;
+        }
+
         notificationSettingRepository.deleteAll(notificationSettingList);
     }
 }
